Update README.md
Please help me with my idea. I am not a programmer at all and really would like to build this cause I think that is could be something that mimics reality some what and would be fun to have a really complex agent built off of a 10 object system, that have a 1024-10 possible different arrangements, meaning an agent is defined by the 10230 points that are not what it is that is passed on via "objectenetical" means. But that is also a way in which these hypercomplex objects can be transformed for our entertainment for these little self governing objectenetical discount "sims" to be plagued with existence too.